+++ src/components/values/MarkupView.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import type MarkupValue from '@values/MarkupValue';
-    import MarkupHTMLView from '../concepts/MarkupHTMLView.svelte';
+    import SymbolView from './SymbolView.svelte';
+    import Sym from '@nodes/Sym';
 
     interface Props {
         value: MarkupValue;
@@ -10,4 +11,4 @@
     let { value, inline = true }: Props = $props();
 </script>
 
-<MarkupHTMLView markup={value.markup} {inline} />
+<SymbolView symbol={value.markup.toWordplay()} type={Sym.Words} />